requirements.txt
requests==2.22.0
google-cloud-bigquery==2.6.1
pandas==1.0.5
pyarrow==2.0.0
google-cloud-bigquery-storage==2.1.0
git+https://github.com/pydata/pandas-gbq.git 